Only start conn once in active state (#128)
We discovered an issue where a conn could get stuck in a pending state if
it the remote peer closed the conn (sent EOF) before the local peer moved
the conn into an active state, because connClosedEvent assumes that the
conn is already in an active state.

By not starting the conn until it is in active state, this diff eliminates the possibility
that it will close itself while still in a pending state.

Note, an underlying closed socket != a closed conn.Conn. conn.Conn is
"closed" only when Close() is explicitly called.

Unfortunately, it is near impossible to test for this race condition given the
distributed nature of scheduler.