Fix shadow backend mode auth config parsing (#362)
**Why**
Currently, the shadow backend mode does not parse the backend credentials correctly. The structure of the Go object with all backend auth credentials looks like this:
```
{
    "s3": {
      <s3 credentials>
    },
   "gcs": {
	<gcs credentials>
    },
}
```

Upon creating a backend client, we look for a key/value pair in this file where the key is the name of the backend we are using (e.g. we would look for the key/value "s3"/<s3-credentials>). 

However, when we use the "shadow" backend, this does not work anymore - Kraken looks for a key/value pair with "shadow" as the key. Instead, Kraken should:
1. Figure out what active and shadow backends it uses (e.g. s3 and gcs).
2. Look for their key/value pairs.

To make this change, we'll change the way we pass the master auth config to the backend client factories - up until this point, we passed only the auth credentials of the respective backend. However, the shadow backend doesn't have its own credentials but rather uses the credentials of two other backends (e.g. gcs and s3). So instead, we'll pass the full config to the `factory.Create` methods and they can parse it as they require.

**What**
 - Properly parse auth credentials when using shadow backend mode.
 - Pass master auth config (the config with the credentials of all backends) to `factory.Create` methods. The Create methods then parse the config as they require.
 - Rename variables for readability.
 - Correct 2 typos in error logs in the SQL backend.