Add tracker readiness endpoint and client (#380)
As described in #371, all services should implement a "readiness" endpoint, which checks whether the service is ready to serve traffic, i.e. whether it can reach all of its dependencies. The tracker makes requests to the origin cluster to get metainfo data for torrents. Therefore, its readiness endpoint should check the origin's readiness.

Additionally, the tracker's readiness endpoint must be queried by the agent, thus a client for the endpoint must be added.

- Add readiness endpoint for tracker, which checks origin's readiness.
- Add client for the endpoint