Remove regex from readiness tests (#383)
The refactored tests assert that when a service's readiness endpoint fails, it returns an error including the URL used to call the endpoint. Currently, the URL is matched against a regex, which is not ideal. This PR replaces the regex with a normal string, which is much easier to read.