feat(lib/backend): add functionality to close clients (#451)
* feat(lib/backend): add functionality to close clients

What?
This PR aims to add functionality to close the clients and it's
resources in case the resources have closers.

Why?
There are use cases in for gcsbacknd using `storage.transfermanager` to
close the single downloader it creates. If we don't close it, there's a
possibility of leaking resources.

How?
Add `Close` to client interface and implement the method for each
client.\
Then `backend.Manager` will use them to expose a `Close()` method that
can be used by build-index and origin to close them.

* add close methods to tests for proper cleanups

* resolved comments